requirements.txt
Flask==0.12.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
requests==2.18.4
Flask_script==2.0.6
Flask_migrate==2.1.1
flask_cache==0.13.1
flask_mail==0.9.1
Flask-RESTful==0.3.6
pillow==5.0.0